requirements.txt
# Core dependencies for inference
google-cloud-aiplatform>=1.70.0
google-cloud-storage>=2.18.0
python-dotenv>=1.0.0

# Scoring app dependencies
streamlit>=1.28.0
pandas>=2.0.0

# Visualization dependencies
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0